+++ tests/zxndma.tests.ts
@@ -12,7 +12,7 @@ suite('ZxnDma', function () {
 	let dma;
 
 	setup(() => {
-		const ports = new Z80Ports(0xFF);
+		const ports = new Z80Ports(true);
 		const memory = new SimulatedMemory(new MemoryModelAllRam(), ports);
 		dma = new ZxnDma(memory, ports) as any;
 	});
@@ -637,7 +637,7 @@ suite('ZxnDma', function () {
 			dma.serialize(mem);
 
 			// Continue with different dma instance
-			const ports2 = new Z80Ports(0xFF);
+			const ports2 = new Z80Ports(true);
 			const memory2 = new SimulatedMemory(new MemoryModelAllRam(), ports2);
 			const dma2 = new ZxnDma(memory2, ports2) as any;
 			dma2.deserialize(mem);